remove v1/assist
