+++ crates/runtime/src/http/v1/mod.rs
@@ -13,7 +13,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-pub mod assist;
 pub mod catalogs;
 pub mod chat;
 pub mod datasets;
