+++ crates/runtime/src/http/v1/search.rs
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 use crate::embeddings::vector_search::{RetrievalLimit, VectorSearch, VectorSearchResult};
+use arrow::array::RecordBatch;
 use axum::{
     http::StatusCode,
     response::{IntoResponse, Response},
@@ -23,8 +24,6 @@ use datafusion::sql::TableReference;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::{collections::HashMap, sync::Arc};
-
-use super::assist::create_primary_key_payload;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "lowercase")]
@@ -110,4 +109,32 @@ pub(crate) async fn post(
             (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response()
         }
     }
+}
+
+#[allow(clippy::from_iter_instead_of_collect)]
+pub(crate) fn create_primary_key_payload(
+    table_primary_keys: &HashMap<TableReference, Vec<RecordBatch>>,
+) -> Result<HashMap<String, Value>, Box<dyn std::error::Error>> {
+    let from_value_iter = table_primary_keys
+        .iter()
+        .map(|(tbl, pks)| {
+            let buf = Vec::new();
+            let mut writer = arrow_json::ArrayWriter::new(buf);
+            for pk in pks {
+                writer.write_batches(&[pk])?;
+            }
+            writer.finish()?;
+            let res: Value = match String::from_utf8(writer.into_inner()) {
+                Ok(res) => serde_json::from_str(&res)?,
+                Err(e) => {
+                    tracing::debug!("Error converting JSON buffer to string: {e}");
+                    serde_json::Value::String(String::new())
+                }
+            };
+            Ok((tbl.to_string(), res))
+        })
+        .collect::<Result<Vec<_>, Box<dyn std::error::Error>>>()?;
+    let from_value: HashMap<String, Value> =
+        HashMap::from_iter(from_value_iter.iter().map(|(k, v)| (k.clone(), v.clone())));
+    Ok(from_value)
 }
